Finish centralization of API models in `mcp-tef-models`
There were some missing models to be added to the new package.
This PR finishes adding the last missing models and correctly
fixes all the imports
