+++ src/mcp_tef/api/mcp_servers.py
@@ -7,17 +7,15 @@ from fastapi import APIRouter, Depends, Query, Request
 from fastapi.security import APIKeyHeader
 from mcp_tef_models.schemas import (
     MCPServerToolsResponse,
+    ToolQualityResponse,
+    ToolQualityResult,
 )
 
 from mcp_tef.api.errors import BadRequestError
 from mcp_tef.config.settings import Settings, get_settings
 from mcp_tef.services.llm_service import LLMService
 from mcp_tef.services.mcp_loader import MCPLoaderService
-from mcp_tef.services.tool_quality_service import (
-    ToolQualityResponse,
-    ToolQualityResult,
-    ToolQualityService,
-)
+from mcp_tef.services.tool_quality_service import ToolQualityService
 
 logger = structlog.get_logger(__name__)
