--- src/mcp_tef/models/enums.py
@@ -1,9 +1,0 @@
-"""Enums for mcp-tef models.
-
-This module is deprecated. New code should import directly from mcp_tef_models.enums.
-"""
-
-# This file is kept for backward compatibility but should not be used in new code.
-# All enums have been moved to mcp_tef_models.enums
-
-__all__: list[str] = []
